requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
pymongo==4.6.1
python-dotenv==1.0.1
bson==0.5.10 